harden Stripe webhook idempotency, refine Elasticsearch indexing, switch price to scaled_float, add real-DB infra tests, fix payment frontend errors

+++ src/CoursePlatform.Application/Features/Courses/Queries/GetCourses/CourseListDto.cs
@@ -20,4 +20,4 @@ public record CourseListDto(
     int ReviewCount,
     IReadOnlyCollection<string> MatchedBy);
 
-public record CoursesVm(IReadOnlyCollection<CourseListDto> Items, int TotalCount);
+public record CoursesVm(IReadOnlyCollection<CourseListDto> Items, long TotalCount);
